Main: allow forcing DXT decompression via Codec::setParameter (#3291)
The DDSCodec keeps the DXT format if the hardware supports it. However now it is possible to enforce the DXT decompression into RGB(A) format. This is useful e.g. when we have DDS texture without mipmaps and we need them. DirectX11 does not generate mipmaps if the texture is compressed (DirectX9 does it automatically).

Also added support to obtain the DDS stream (not only saving to file).

Co-authored-by: stupak <stupak@ekosoft.sk>