Main: use constexpr for static Vector constants (#3315)
This fixes the initialization-order-fiasco for static
Vector constants while preserving API (though breaking
ABI) through the following steps:

* Make constexpr the VectorBase and Vector constructors
* Instantiate global constexpr variables in a hidden
  namespace in .cpp file
* Change static class constants to references (ABI break)
  and assign to global constexpr variables

This matches the approach taken in #3313. Note that
the Vector(T) constructor could not be made constexpr,
so the ZERO and SCALE variables were constructed with
repeated arguments instead of a single argument.